backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
azure-functions
azure-cosmos
azure-communication-chat>=1.2.0
azure-communication-sms>=1.0.0
azure-ai-contentsafety>=1.0.0
azure-search-documents>=11.4.0
azure-identity>=1.14.0
azure-keyvault-secrets>=4.7.0
openai>=1.0.0
aiohttp
